# combi-uees-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMBI UEES con Laravel 12",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rutas/{ruta}/motoristas": {
            "get": {
                "tags": [
                    "Motoristas"
                ],
                "summary": "Obtener todos los motoristas de una ruta",
                "operationId": "c71236c8ecb2982f4d47140489e73c00",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de motoristas ordenados por creación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Motorista"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Motoristas"
                ],
                "summary": "Registrar un nuevo motorista en una ruta",
                "operationId": "579b9995ba8dd29becd5787633c0f49a",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "motoristaURLFotoDePerfil",
                                    "motoristaEstado",
                                    "motoristaUbicación",
                                    "IDRuta"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Carlos López"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "motorista@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contraseñaSegura123"
                                    },
                                    "motoristaURLFotoDePerfil": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://ejemplo.com/foto.jpg"
                                    },
                                    "motoristaEstado": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "motoristaUbicación": {
                                        "type": "string",
                                        "example": "UEES"
                                    },
                                    "IDRuta": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Motorista registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Motorista"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/motoristas/{user}": {
            "put": {
                "tags": [
                    "Motoristas"
                ],
                "summary": "Actualizar los datos de un motorista",
                "operationId": "2d82c8795d19f356a25d0ee00cac12ba",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del motorista",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "password",
                                    "motoristaURLFotoDePerfil",
                                    "motoristaEstado",
                                    "motoristaUbicación",
                                    "IDRuta"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Carlos López"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nuevaContraseñaSegura123"
                                    },
                                    "motoristaURLFotoDePerfil": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://ejemplo.com/nueva-foto.jpg"
                                    },
                                    "motoristaEstado": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "motoristaUbicación": {
                                        "type": "string",
                                        "example": "Ruta Norte"
                                    },
                                    "IDRuta": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Motorista actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Motorista"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/rutas/{ruta}/motoristas/{user}/estado": {
            "patch": {
                "tags": [
                    "Motoristas"
                ],
                "summary": "Cambiar el estado de un motorista",
                "operationId": "3dfe2a4981cb085c2abc2e5134181b26",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del motorista",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motoristaEstado"
                                ],
                                "properties": {
                                    "motoristaEstado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado del motorista actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Motorista"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/rutas/{ruta}/motoristas/{user}": {
            "delete": {
                "tags": [
                    "Motoristas"
                ],
                "summary": "Eliminar un motorista",
                "operationId": "2b903332832418d10e8a589c78286741",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del motorista",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Motorista eliminado exitosamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/usuario": {
            "get": {
                "tags": [
                    "Motoristas"
                ],
                "summary": "Obtener información del usuario autenticado",
                "operationId": "491ad9f6d295f554e7ae03e37222e2ac",
                "responses": {
                    "200": {
                        "description": "Información del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Motorista"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Motorista": {
                "title": "Motorista",
                "description": "Modelo de un motorista asignado a una ruta",
                "required": [
                    "id",
                    "name",
                    "email",
                    "motoristaURLFotoDePerfil",
                    "motoristaEstado",
                    "motoristaUbicación",
                    "IDRuta"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Carlos López"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "motorista@example.com"
                    },
                    "motoristaURLFotoDePerfil": {
                        "type": "string",
                        "format": "url",
                        "example": "https://ejemplo.com/foto.jpg"
                    },
                    "motoristaEstado": {
                        "type": "boolean",
                        "example": true
                    },
                    "motoristaUbicación": {
                        "type": "string",
                        "example": "Campus UEES"
                    },
                    "IDRuta": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-26T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-26T12:45:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}